Bean Fix
An error suddenly occurred stating that bean metadataExtractor's property extraMetadataFieldExtractors required one bean and was supplied 3. It was recommended to mark one as primary, which is exactly what I did to whiteBoxViewMetadataNames. There was no reasoning behind this decision, but it did fix the problem. I do not know what suddenly caused this problem to occur.